Extract PayPalViewModel business logic into PayPalFragment.
