+++ Demo/build.gradle
@@ -14,7 +14,9 @@ android {
 
     defaultConfig {
         applicationId "com.paypal.android"
-        minSdkVersion 21
+        // TODO: revert once Native XO supports minSdk 21
+        //minSdkVersion 21
+        minSdkVersion 23
         targetSdkVersion 30
         versionCode 1
         versionName "1.0"
